# api_v2_backup_20251016_203324/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_EduardoMG12_cine_api_v2_internal_dto.AuthResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.UserProfile'
    type: object
  github_com_EduardoMG12_cine_api_v2_internal_dto.ConfirmEmailRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
  github_com_EduardoMG12_cine_api_v2_internal_dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  github_com_EduardoMG12_cine_api_v2_internal_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_EduardoMG12_cine_api_v2_internal_dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  github_com_EduardoMG12_cine_api_v2_internal_dto.RegisterRequest:
    properties:
      display_name:
        maxLength: 100
        minLength: 2
        type: string
      email:
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - display_name
    - email
    - password
    - username
    type: object
  github_com_EduardoMG12_cine_api_v2_internal_dto.ResetPasswordRequest:
    properties:
      new_password:
        maxLength: 100
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
  github_com_EduardoMG12_cine_api_v2_internal_dto.UserProfile:
    properties:
      bio:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: integer
      is_private:
        type: boolean
      profile_picture_url:
        type: string
      theme:
        type: string
      username:
        type: string
    type: object
  github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse:
    properties:
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cineverse.com
    name: CineVerse API Support
    url: http://www.cineverse.com/support
  description: A comprehensive social network platform for movie enthusiasts
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://cineverse.com/terms/
  title: CineVerse API
  version: "2.0"
paths:
  /auth/confirm-email:
    post:
      consumes:
      - application/json
      description: Verify user's email address using confirmation token
      parameters:
      - description: Email confirmation token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.ConfirmEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email confirmed successfully
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.MessageResponse'
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
      summary: Confirm email address
      tags:
      - Authentication
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset link to user's email
      parameters:
      - description: User email for password reset
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset email sent
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.MessageResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
      summary: Request password reset
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user credentials and return access token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.AuthResponse'
        "400":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
      summary: User login
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email verification
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.AuthResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using reset token from email
      parameters:
      - description: Password reset token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_dto.MessageResponse'
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_EduardoMG12_cine_api_v2_internal_utils.ErrorResponse'
      summary: Reset user password
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"